requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=1.12.0
torchvision>=0.13.0
timm>=0.9.0
Pillow>=9.0.0
numpy>=1.21.0